# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sessio API",
        "description": "API para gerenciamento de sessões, clientes e bloqueios de tempo.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/events": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Listar todas as sessões do usuário autenticado",
                "operationId": "7058ab7c89d20c752b6afd6d808f6526",
                "responses": {
                    "200": {
                        "description": "Lista de sessões",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Event"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Criar nova sessão",
                "operationId": "508895e03d7f10edac84357084c87b20",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sessão criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflito de horário"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Exibir uma sessão específica",
                "operationId": "6f743f71f7a88dbe57be1d9c9ebf47cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da sessão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da sessão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "404": {
                        "description": "Sessão não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Atualizar uma sessão",
                "operationId": "fb130dcf1ab4d7abbd3a173741df1f8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da sessão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sessão atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "404": {
                        "description": "Sessão não encontrada"
                    },
                    "409": {
                        "description": "Conflito de horário"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Excluir uma sessão",
                "operationId": "b8c8f4b41a32532f5e9f7146e511bc61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da sessão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sessão excluída com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sessão excluída com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "404": {
                        "description": "Sessão não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/calendar-items": {
            "get": {
                "tags": [
                    "Calendar"
                ],
                "summary": "Listar todos os itens do calendário (sessões e bloqueios) em um período",
                "operationId": "fadf19ba42f735ddf6195f0ec7612b78",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Data de início do período (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "Data de término do período (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de itens do calendário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "item_type": {
                                                "type": "string",
                                                "enum": [
                                                    "session",
                                                    "block"
                                                ]
                                            }
                                        },
                                        "type": "object",
                                        "oneOf": [
                                            {
                                                "$ref": "#/components/schemas/Event"
                                            },
                                            {
                                                "$ref": "#/components/schemas/TimeBlock"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Event": {
                "title": "Event",
                "description": "Event model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "client_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-08T14:00:00Z"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-08T15:00:00Z"
                    },
                    "duration_min": {
                        "type": "integer",
                        "example": 60
                    },
                    "focus_topic": {
                        "type": "string",
                        "example": "Ansiedade e estresse no trabalho",
                        "nullable": true
                    },
                    "session_notes": {
                        "type": "string",
                        "example": "Paciente relatou melhora nos sintomas de ansiedade após as técnicas de respiração.",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "In-person",
                            "Online"
                        ],
                        "example": "In-person"
                    },
                    "payment_status": {
                        "type": "string",
                        "enum": [
                            "Pending",
                            "Paid"
                        ],
                        "example": "Pending"
                    },
                    "session_status": {
                        "type": "string",
                        "enum": [
                            "Scheduled",
                            "Completed",
                            "Canceled"
                        ],
                        "example": "Scheduled"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-08T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-08T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "EventRequest": {
                "title": "Event Request",
                "description": "Request payload for creating/updating an event",
                "required": [
                    "client_id",
                    "start_time",
                    "duration_min",
                    "type",
                    "payment_status"
                ],
                "properties": {
                    "client_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-08T14:00:00Z"
                    },
                    "duration_min": {
                        "type": "integer",
                        "maximum": 1440,
                        "minimum": 1,
                        "example": 60
                    },
                    "focus_topic": {
                        "type": "string",
                        "example": "Ansiedade e estresse no trabalho",
                        "nullable": true
                    },
                    "session_notes": {
                        "type": "string",
                        "example": "Paciente relatou melhora nos sintomas de ansiedade após as técnicas de respiração.",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "In-person",
                            "Online"
                        ],
                        "example": "In-person"
                    },
                    "payment_status": {
                        "type": "string",
                        "enum": [
                            "Pending",
                            "Paid"
                        ],
                        "example": "Pending"
                    },
                    "session_status": {
                        "type": "string",
                        "enum": [
                            "Scheduled",
                            "Completed",
                            "Canceled"
                        ],
                        "example": "Scheduled"
                    }
                },
                "type": "object"
            },
            "TimeBlock": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Almoço"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00-03:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T13:00:00-03:00"
                    },
                    "color_hex": {
                        "type": "string",
                        "maxLength": 7,
                        "example": "#FF5733",
                        "nullable": true
                    },
                    "emoji": {
                        "type": "string",
                        "maxLength": 5,
                        "example": "🍽️",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "item_type": {
                        "type": "string",
                        "example": "block"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}